Deprecate Proxy.Authorization in favor of top-level ProxyAuthorization
Proxy.Authorization never referenced Proxy's Headers generic parameter,
but being declared as a struct nested inside Proxy<Headers> meant
Proxy<A>.Authorization and Proxy<B>.Authorization were formally
different types to the compiler. That's a trap for any code
referencing the type outside a context where Headers is already
pinned by argument inference (e.g. a standalone helper function),
discovered while implementing swift-configuration support.

ProxyAuthorization is now a plain top-level type; Proxy.Authorization
becomes a deprecated typealias to it, which keeps existing call sites
compiling (typealiases don't introduce new nominal types, so the
alias resolves to the same concrete type regardless of Headers).
